Resolve circular import between models and telegrambot
