+++ teleteam_bot/telegrambot.py
@@ -1,4 +1,5 @@
 import os
+import logging
 
 from telegram.ext import (
     CommandHandler,
@@ -10,11 +11,8 @@ from telegram.ext import (
 )
 from telegram import ParseMode, InlineKeyboardMarkup, InlineKeyboardButton
 from django_telegrambot.apps import DjangoTelegramBot
-from django.core.files import File
 
 
-
-import logging
 
 from .user_commands.create_task import CREATE_TASK_HANDLER
 from .user_commands.help import help
@@ -67,35 +65,3 @@ def main():
     dp.add_error_handler(error)
 
     __bot = dp.bot
-
-
-def get_group_photo(group_chat_id):
-    # Try to get the telegram chat photo
-    try:
-        # Get chat object from Telegram API
-        chat = __bot.get_chat(group_chat_id)
-
-        # Get photo url from Telegram API
-        photo_file = chat.photo.get_small_file()
-
-        path = settings.MEDIA_ROOT+str(chat.id)+'.jpg'
-
-        # Download to media folder
-        photo_file.download(custom_path=path)
-
-        # Open the file
-        photo = File(open(path, 'rb'))
-
-        # Remove the temporary file path
-        os.remove(path)
-
-        print(f'Group Photo is retrieved for {chat.title}')
-
-        return File(photo)
-    except Exception as e:
-        print(e)
-        return None
-
-
-def get_bot():
-    return __bot
